Updated tests to check for json schema equality
