+++ instructor/__init__.py
@@ -21,7 +21,7 @@ from .client import (
     from_litellm,
     Provider,
 )
-
+from .validators import async_field_validator, async_model_validator
 
 __all__ = [
     "Instructor",
@@ -44,6 +44,8 @@ __all__ = [
     "Instructions",
     "handle_parallel_model",
     "handle_response_model",
+    "async_field_validator",
+    "async_model_validator",
 ]
