+++ tests/llm/test_openai/test_async_decorator.py
@@ -5,8 +5,10 @@ import instructor
 from openai import AsyncOpenAI
 from instructor import from_openai
 from .util import models, modes
-from instructor.decorators import async_field_validator, async_model_validator
+from instructor import async_field_validator, async_model_validator
+from instructor.function_calls import openai_schema
 from instructor.function_calls import OpenAISchema
+from pydantic import BaseModel, Field
 
 
 class UserExtractValidated(OpenAISchema):
@@ -24,7 +26,6 @@ class UserExtractValidated(OpenAISchema):
 
 @pytest.mark.parametrize("model, mode", product(models, modes))
 @pytest.mark.asyncio
-@pytest.mark.skip()
 async def test_simple_validator(model, mode, aclient):
     aclient = instructor.from_openai(aclient, mode=mode)
     model = await aclient.chat.completions.create(
@@ -78,7 +79,6 @@ class ExtractedContent(OpenAISchema):
 
 @pytest.mark.parametrize("model, mode", product(models, modes))
 @pytest.mark.asyncio
-@pytest.mark.skip()
 async def test_async_validator(model, mode, aclient):
     aclient = instructor.from_openai(aclient, mode=mode)
     content = """
@@ -115,7 +115,6 @@ async def test_async_validator(model, mode, aclient):
 
 @pytest.mark.parametrize("model, mode", product(models, modes))
 @pytest.mark.asyncio
-@pytest.mark.skip()
 async def test_nested_model(model, mode, aclient):
     class Users(OpenAISchema):
         users: list[UserExtractValidated]
@@ -240,3 +239,63 @@ async def test_context_passing_in_nested_field_validator():
         str(res[0])
         == "Exception of Invalid Error but with {'abcdef': '123'}! encountered at model.user_names"
     )
+
+
+class User(BaseModel):
+    name: str = Field(description="User's name")
+    age: int = Field(description="User's age")
+    email: str = Field(description="User's email address")
+
+
+def test_openai_schema_serialization():
+    UserSchema = openai_schema(User)
+    assert User.model_json_schema() == UserSchema.model_json_schema()
+
+
+def test_nested_class():
+    class Users(BaseModel):
+        users: list[User]
+        user: User
+
+    assert Users.model_json_schema() == openai_schema(Users).model_json_schema()
+
+
+def test_nested_class_with_async_decorators():
+    class NestedUserWithValidation(BaseModel):
+        name: str
+
+        @async_field_validator("name")
+        async def validate_uppercase(self, v: str) -> str:
+            await asyncio.sleep(2)
+            return v
+
+    class Users(BaseModel):
+        users: list[NestedUserWithValidation]
+
+    assert Users.model_json_schema() == openai_schema(Users).model_json_schema()
+
+
+def test_nested_class_with_multiple_async_decorators():
+    class User(BaseModel):
+        name: str
+        age: int
+
+        @async_model_validator()
+        async def validate_user(self) -> "User":
+            await asyncio.sleep(1)
+            if self.age < 0:
+                raise ValueError("Age must be non-negative")
+            return self
+
+    class Users(BaseModel):
+        users: list[User]
+        user: User
+
+        @async_model_validator()
+        async def validate_users(self) -> "Users":
+            await asyncio.sleep(1)
+            if len(self.users) == 0:
+                raise ValueError("Users list cannot be empty")
+            return self
+
+    assert Users.model_json_schema() == openai_schema(Users).model_json_schema()
